Finally updated the code to build against the latest versions of libnds, libfat, and wifilib.  Many of these changes were renames to use the new register names, however some changes had to be made to accomodate the new initialization code present in libnds.
Additionally, the code has been fixed such that cross-compiling NHDS on
a 64-bit host works correctly (NetHack uses long's in a few places where
it really matters that the host and target platforms end up with the
same widths.  As a result, that code had to be changed to use uint32_t).

Finally, in the process of fixing some keyboard related bugs, I
reformatted and substantially cleaned up the keyboard code so it's
actually readable and comprehensible, as well as clean and consistent.

+++ include/global.h
@@ -6,6 +6,7 @@
 #define GLOBAL_H
 
 #include <stdio.h>
+#include <stdint.h>
 
 
 /* #define BETA	*/	/* if a beta-test copy	[MRS] */
@@ -302,10 +303,10 @@ extern long *FDECL(alloc, (unsigned int));		/* alloc.c */
 /* Used for consistency checks of various data files; declare it here so
    that utility programs which include config.h but not hack.h can see it. */
 struct version_info {
-	unsigned long	incarnation;	/* actual version number */
-	unsigned long	feature_set;	/* bitmask of config settings */
-	unsigned long	entity_count;	/* # of monsters and objects */
-	unsigned long	struct_sizes;	/* size of key structs */
+	uint32_t	incarnation;	/* actual version number */
+	uint32_t	feature_set;	/* bitmask of config settings */
+	uint32_t	entity_count;	/* # of monsters and objects */
+	uint32_t	struct_sizes;	/* size of key structs */
 };
